+++ squel-basic.js
@@ -235,6 +235,10 @@ OTHER DEALINGS IN THE SOFTWARE.
       return item;
     };
 
+    BaseBuilder.prototype._escapeValue = function(str) {
+      return str;
+    };
+
     BaseBuilder.prototype._formatValue = function(value) {
       var customHandler;
       customHandler = getValueHandler(value, this.options.valueHandlers, cls.globalValueHandlers);
@@ -247,6 +251,7 @@ OTHER DEALINGS IN THE SOFTWARE.
         value = value ? "TRUE" : "FALSE";
       } else if ("number" !== typeof value) {
         if (false === this.options.usingValuePlaceholders) {
+          value = this._escapeValue(value);
           value = "'" + value + "'";
         }
       }